feat: implement Sequelize data models for core school management entities

+++ backend_node/src/models/Attendance.js
@@ -4,8 +4,8 @@ const sequelize = require('../config/db');
 const Attendance = sequelize.define('Attendance', {
   id: { type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true },
   school_id: { type: DataTypes.BIGINT, allowNull: false },
-  student_id: { type: DataTypes.INTEGER, allowNull: false },
-  classroom_id: { type: DataTypes.INTEGER },
+  student_id: { type: DataTypes.BIGINT, allowNull: false },
+  classroom_id: { type: DataTypes.BIGINT },
   date: { type: DataTypes.DATE, allowNull: false },
   status: { type: DataTypes.STRING, defaultValue: 'present' }, // present, absent, late, excused
   remarks: { type: DataTypes.TEXT },
