+++ backend_node/src/models/Class.js
@@ -7,9 +7,9 @@ const Class = sequelize.define('Class', {
   name: { type: DataTypes.STRING, allowNull: false },
   form: { type: DataTypes.STRING }, // e.g., "Form 1", "Grade 10"
   category: { type: DataTypes.STRING }, // e.g., "Science", "Arts", "Commercial"
-  class_teacher_id: { type: DataTypes.INTEGER },
+  class_teacher_id: { type: DataTypes.BIGINT },
   capacity: { type: DataTypes.INTEGER },
-  academic_year_id: { type: DataTypes.INTEGER },
+  academic_year_id: { type: DataTypes.BIGINT },
   is_active: { type: DataTypes.BOOLEAN, defaultValue: true },
   created_at: { type: DataTypes.DATE, defaultValue: DataTypes.NOW },
 }, {
