+++ backend_node/src/models/Grade.js
@@ -4,10 +4,10 @@ const sequelize = require('../config/db');
 const Grade = sequelize.define('Grade', {
   id: { type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true },
   school_id: { type: DataTypes.BIGINT, allowNull: false },
-  student_id: { type: DataTypes.INTEGER, allowNull: false },
-  subject_id: { type: DataTypes.INTEGER, allowNull: false },
-  term_id: { type: DataTypes.INTEGER, allowNull: false },
-  classroom_id: { type: DataTypes.INTEGER },
+  student_id: { type: DataTypes.BIGINT, allowNull: false },
+  subject_id: { type: DataTypes.BIGINT, allowNull: false },
+  term_id: { type: DataTypes.BIGINT, allowNull: false },
+  classroom_id: { type: DataTypes.BIGINT },
   ca: { type: DataTypes.FLOAT }, // Continuous Assessment
   midterm: { type: DataTypes.FLOAT },
   final: { type: DataTypes.FLOAT },
