+++ backend_node/src/models/Term.js
@@ -4,7 +4,7 @@ const sequelize = require('../config/db');
 const Term = sequelize.define('Term', {
   id: { type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true },
   school_id: { type: DataTypes.BIGINT, allowNull: false },
-  academic_year_id: { type: DataTypes.INTEGER, allowNull: false },
+  academic_year_id: { type: DataTypes.BIGINT, allowNull: false },
   name: { type: DataTypes.STRING, allowNull: false }, // e.g., "Term 1", "Semester 1"
   start_date: { type: DataTypes.DATE },
   end_date: { type: DataTypes.DATE },
